refactor: DELETE /{productId} 상품삭제기능 리펙토링
- 사용자 권한체크 추가
- 사용하지 않는 불필요한 코드 제거

+++ src/main/java/org/mwt/market/domain/product/service/ProductService.java
@@ -77,8 +77,12 @@ public class ProductService {
     }
 
     @Transactional
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(Long productId, UserPrincipal userPrincipal) {
         Product product = productRepository.findById(productId).orElseThrow(AlreadyGoneException::new);
+        if (!userPrincipal.getId().equals(product.getSeller().getUserId())) {
+            throw new NoPermissionException();
+        }
+
         product.delete();
         productRepository.save(product);
     }
